Move new test classes to ch.astorm.jotlmsg.io.
